+++ CHANGES.md
@@ -11,6 +11,11 @@ E.g. to retrieve the changes made between release 3.4 and release 3.5, type::
 The contributors lists are in no order and apologies to those accidentally not
 mentioned. If we missed you, please let us know!
 
+CHANGES  4.3.2a2 07/03/2025
+---------------------------
+    * table styles have 'DIRECTION' for RTL/LTR etc
+    * now using rlbidi with bidiWordList for single frag cases, splitting works
+
 CHANGES  4.3.2a1 25/02/2025
 ---------------------------
     * fix most single frag RTL/LTR para measurements